value-emitter => events package
