+++ packages/core/logger/src/Logger.js
@@ -2,7 +2,7 @@
 
 import type {IDisposable, LogEvent} from '@parcel/types';
 
-import ValueEmitter from '@parcel/value-emitter';
+import {ValueEmitter} from '@parcel/events';
 
 class Logger {
   #logEmitter = new ValueEmitter<LogEvent>();
