+++ packages/utils/events/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "@parcel/value-emitter",
+  "name": "@parcel/events",
   "version": "2.0.0",
   "license": "MIT",
   "repository": {
     "type": "git",
     "url": "https://github.com/parcel-bundler/parcel.git"
   },
-  "main": "src/ValueEmitter.js",
+  "main": "src/index.js",
   "scripts": {
     "build": "babel src -d lib",
     "prepublish": "yarn build",
