+++ packages/utils/events/src/errors.js
@@ -1,0 +1,6 @@
+// @flow strict-local
+
+export class AlreadyDisposedError extends Error {}
+Object.defineProperty(AlreadyDisposedError.prototype, 'name', {
+  value: 'AlreadyDisposedError'
+});
