+++ packages/utils/events/src/index.js
@@ -1,0 +1,4 @@
+// @flow strict-local
+
+export {default as ValueEmitter} from './ValueEmitter';
+export {AlreadyDisposedError} from './errors';
