+++ packages/utils/events/src/types.js
@@ -1,0 +1,5 @@
+// @flow strict-local
+
+export interface IDisposable {
+  dispose(): void;
+}
