+++ packages/utils/events/test/.eslintrc
